requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
pandas
openpyxl==3.1.2
WeasyPrint==60.2
python-dotenv==1.0.0
gunicorn==21.2.0
email_validator
